quicklogic: add slow and fast corners architectures
This commit adds two different devices and boards for two different
corners.

The CMake infrastructure does not allow for a same device to use different arch
XML and RR graphs (that have different corner data).

The multiple corner testing can be properly added to the installation
package testing suite (which still needs to be developed) so that we
can allow and test usage of different corners for the same device during
different parts of the flow (e.g. SLOW corner for PnR and FAST corner
for analysis)

Signed-off-by: Alessandro Comodi <acomodi@antmicro.com>