ci: tests: use one job for all_route_tests
There are some unrelated test cases which fail if multiple jobs are
used. Unclear what is causing the problem at the moment.

For now, only one job is used when performing the route tests. Given
that these are quite small tests for dummy/small architectures, the
impact on CI run-time is acceptable.

Signed-off-by: Alessandro Comodi <acomodi@antmicro.com>